v2 collisions + autoSync and autoRemoval behaviors
autoRemoval will likely be refined to "autoAdopt."  It helps solve the
issue of local clients disconnecting from a session but leaving objects
in the Firebase session that they are still responsible for the logic
of.  After a timeout, all other users in the room request "ownership" of
the object, and it is granted to one individual client a couple seconds
later.  Currently, that client just removes the object right away, but
this can easily be modified to allow the app to specify a callback for
when this event occurs.